% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trips.R
\name{add_trips}
\alias{add_trips}
\title{Additional trips for trip data set}
\usage{
add_trips(
  trip_ids = 0,
  new_mode = "pedestrian",
  distance = 1,
  participant_id = 0,
  age = 20,
  sex = "male",
  nTrips = 3,
  speed = 4.8
)
}
\arguments{
\item{trip_ids}{ids for new trips}

\item{new_mode}{mode for new trips}

\item{distance}{distances to sample from}

\item{participant_id}{participant id for new trips}

\item{age}{age for participant}

\item{sex}{sex for participant}

\item{nTrips}{number of trips for each participant}

\item{speed}{speed for new trips}
}
\value{
data frame of trips
}
\description{
Creates a data frame with given characteristics to be added to an existing trip data set.
}
\details{
This function is used to create new trips with certain characteristics that can be added to an existing trip data set.
The input distance given is the upper limit of trip distances added and the function takes the number of trips per
person (nTrips) random samples between 1 and the input distance given. Age and sex of each participant are also
sampled from a range of input age and sex values.

The function performs the following steps:
\itemize{
\item create nTrips new trips sampling from distances, ages and sexes
}
}
