% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AP_dose_response.R
\name{AP_dose_response}
\alias{AP_dose_response}
\title{Calculate relative risk given PM exposure level}
\usage{
AP_dose_response(cause, dose, quantile, confidence_intervals = F)
}
\arguments{
\item{cause}{name of disease}

\item{dose}{vector of PM exposure levels from individuals for a given age range and scenario}

\item{quantile}{quantile of the dose response functions to be used as output value}

\item{confidence_intervals}{logic: whether or not to return confidence intervals}
}
\value{
data frame of relative risks for each person in the synthetic population and each AP related disease in each scenario
}
\description{
Calculate the relative risk (RR) for each person in the synthetic population for each 
disease related to air pollution and each scenario based on the individual PM exposure levels
}
\details{
This function performs the following steps:

\itemize{
\item various checks to ensure the correct PM exposure levels (doses) and disease are read in

\item get the lookup table for the required dose response functions which contains for each dose
  the median RR and the upper and lower confidence interval RR values

\item for the PM exposure doses in the synthetic population find the needed RR by extrapolating the 
  dose responses given in the lookup table

\item if a confidence interval is required, repeat this interpolation using the upper and lower
  confidence RR values

\item if the required quantile is not 0.5, i.e. the median, then find the required RR value by
  defining a normal function with mean the median RR value for that dose and standard deviations
  defined by the upper and lower confidence RR values. Use the correct quantile from this
  normal function as your RR value.  
}
}
