% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ap_rr.R
\name{ap_dose_response_curve}
\alias{ap_dose_response_curve}
\title{Computes RR as a DR relationship}
\usage{
ap_dose_response_curve(pm, alpha, beta, gamma, tmrel)
}
\arguments{
\item{pm}{PM2.5 exposure}

\item{alpha}{DR parameter}

\item{beta}{DR parameter}

\item{gamma}{DR parameter}

\item{tmrel}{DR parameter}
}
\value{
RR
}
\description{
Computes RR as a DR relationship given four parameters and the PM2.5 exposure
}
