% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_age_groups.R
\name{assign_age_groups}
\alias{assign_age_groups}
\title{Assign age groups to individuals}
\usage{
assign_age_groups(
  dataset,
  age_category = AGE_CATEGORY,
  age_lower_bounds = AGE_LOWER_BOUNDS,
  max_age = MAX_AGE,
  min_age = AGE_LOWER_BOUNDS[1],
  age_label = "age"
)
}
\arguments{
\item{dataset}{data frame to which age categories are to be added and ages outside model scope are removed}

\item{age_category}{vector of strings giving age categories}

\item{age_lower_bounds}{lower boundaries of age categories}

\item{max_age}{maximum age for model}

\item{min_age}{minimum age for model}

\item{age_label}{string label for age column}
}
\value{
edited data frame
}
\description{
Prunes the dataset given max and min ages, i.e. removes entries of people with ages outside the scope of the model and
assigns age group labels given age
}
\details{
The function contains the following steps:
\itemize{
\item remove any dataset entries with ages above or below the max and min ages or with missing age information

\item add age categories
}
}
