% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_synth_pop.R
\name{beta_pointiness}
\alias{beta_pointiness}
\title{Parametrise confidence in PA data}
\usage{
beta_pointiness(confidence)
}
\arguments{
\item{confidence}{value between 0 and 1 representing how confident we are about the PA dataset}
}
\value{
a value to parametrise a beta distribution
}
\description{
Takes a confidence value between 0 and 1 and returns a parameter for a beta distribution
}
