% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_evppi.R
\name{compute_evppi}
\alias{compute_evppi}
\title{Function to calculate the EVPPI values for given input parameters and outcome values}
\usage{
compute_evppi(
  p,
  global_para,
  city_para,
  city_outcomes,
  nsamples,
  individual_para = TRUE
)
}
\arguments{
\item{p}{input parameter index}

\item{global_para}{list of global input parameters that are the same across all cities}

\item{city_para}{list of city specific input parameters}

\item{nsamples}{number of samples}

\item{individual_para}{whether each parameter is to be considered individually or not}

\item{city_outcome}{list of outcomes for a specific city}
}
\value{
list of of EVPPI values for specific city
}
\description{
This function calculates the expected value of partially perfect information, i.e. the percentage by which we could
reduce the variance of the final outcome if we knew this one parameter (or several interdependent parameters) perfectly.
}
\details{
The function performs the following steps:

\itemize{
\item extract the parameter / parameters of interest for which the EVPPI value is to be calculated

\item loop through the various outcomes :
 \itemize{
   \item calculate the EVPPI value for each outcome using the evppivar() function from https://github.com/chjackson/voi
   \item calculate the percentage of variance in the outcome we could reduce were we to know the 
         input parameter / parameters perfectly.
   } 
}
}
