% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_all_scenarios.R
\name{create_all_scenarios}
\alias{create_all_scenarios}
\title{Creates specific scenarios for Accra and Sao Paulo}
\usage{
create_all_scenarios(trip_set)
}
\arguments{
\item{trip_set}{data frame of baseline trips}
}
\value{
list of scenarios
}
\description{
Creates five prespecified scenarios from the baseline for Accra and Sao Paulo
}
