% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_synth_pop.R
\name{create_base_pop}
\alias{create_base_pop}
\title{Creates baseline population}
\usage{
create_base_pop(raw_trip_set)
}
\arguments{
\item{raw_trip_set}{data frame of raw trips taken, bus_driver, new motorcycle and truck trips have already been added}
}
\value{
the baseline population and the trip set which has been pruned and also the proportion of people with non non-occupational

mMET values for each age and sex category
}
\description{
Creates a baseline population by matching individuals in the trip set
to individuals in the physical activity (PA) dataset
}
\details{
The function performs the following steps:

\itemize{
\item adds age category to trip and physical activity datasets by calling assign_age_groups.R

\item To match people in trip data with people in the physical activity dataset:
  \itemize{
  \item create a baseline population by taking the unique participant ids together with age and gender
    information from the trip data (not including bus driver, truck, car driver and commercial
    motorcycle trips)

  \item to assign non-occupational physical activity MMET values to this
    baseline population, the following steps are performed:

   \itemize{
    \item for each sex and age category, find the proportion of people with zero non-occupational MMET values
      and also find the list of people with non-zero MMET values

    \item if BACKGROUND_PA_CONFIDENCE < 1 when calling the value of information script, i.e. when
      input values are sampled from distributions, a beta distribution is built from which the
      proportion of people with zero work and leisure MMET values is sampled using the
      known proportion as mean of this distribution

    \item sample with replacement from a vector with 0 MMET values and the vector non-zero MMET values (from
      the people having non-zero work and leisure MMET values) using the proportion of people with
      zero work and leisure MMET values and assign those sampled MMET values to the baseline population

     }
  }
\item remove participants with trip or stage modes that are not in Vehicle inventory
}
}
