% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cycle_scenarios.R
\name{create_cycle_scenarios}
\alias{create_cycle_scenarios}
\title{Creates cycling scenarios}
\usage{
create_cycle_scenarios(trip_set)
}
\arguments{
\item{trip_set}{data frame of baseline trips}
}
\value{
list of scenarios
}
\description{
Creates five scenarios with 10-50% cycling
}
