% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_max_mode_share_scenarios.R
\name{create_max_mode_share_scenarios}
\alias{create_max_mode_share_scenarios}
\title{Create scenarios defined by maximum mode share}
\usage{
create_max_mode_share_scenarios(trip_set)
}
\arguments{
\item{trip_set}{data frame, baseline scenario}
}
\value{
list of baseline scenario and five mode scenarios
}
\description{
Creates five scenarios where, in each one, the mode share is elevated to the maximum observed across the cities.
The scenario-modes are pedestrian, cycle, car, motorcycle and bus
}
