% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_walk_scenario.R
\name{create_walk_scenario}
\alias{create_walk_scenario}
\title{Create basic pedestrian scenario}
\usage{
create_walk_scenario(trip_set)
}
\arguments{
\item{trip_set}{data frame, baseline scenario}
}
\value{
list of baseline scenario and pedestrian scenario
}
\description{
Duplicate baseline scenario; give each person one 1km walk in the scenario
}
