% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ithim_setup_parameters.R
\name{dirichlet_pointiness}
\alias{dirichlet_pointiness}
\title{Function for Dirichlet parameters}
\usage{
dirichlet_pointiness(confidence)
}
\arguments{
\item{confidence}{value between 0 and 1}
}
\value{
parameterisation
}
\description{
Function to map a confidence value to a parameterisation of a Dirichlet distribution.
}
\details{
Note that the parameterisation is somewhat arbitrary but seems to work on visual inspection.
}
