% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_for_injury_function.R
\name{distances_for_injury_function}
\alias{distances_for_injury_function}
\title{Get distances and model for injuries module}
\usage{
distances_for_injury_function(journeys, dist)
}
\arguments{
\item{journeys}{data frame with total distance (by total population) for each age and sex category and for each scenarios}

\item{dist}{table of (total) distances per mode}
}
\value{
true_distances (mode distances by age and sex with all walking modes and all car modes combined and bus drivers added where relevant),

injuries_list (list of all strike, casualty, age, sex and mode distance combinations for baseline and all scenarios),

reg_model (parameterised Poisson regression model),

injuries_for_model (baseline data containing injury counts for all casualty and strike mode combinations with associated distance data)
}
\description{
Computes exposures (distances) to parameterise the injury regression model, 
which is computed as a Poisson with various offsets and used later in prediction
}
\details{
The function uses distance data and the individual injury tables to perform the following steps
to parameterise the Poisson injury regression model:

 - stage modes aggregated such that all walk related stages (walk and walk to pt) are of the same mode (walk), 
   similarly for all car related journey

 - bus drivers are added to bus journeys (where relevant)

 - Takes Baseline injury tables (whw and nov) and adds distances for each strike and casualty
   mode. Distances are added by age and gender category if there exists such information for 
   the injury counts (injuries_for_model). If there exists a fatality for some casualty and 
   strike mode and age and gender category are aggregated by strike and casualty mode. 
   If, after the aggregation there still exist fatalities for which either casualty 
   or strike mode distance is missing, these fatalities are removed as we cannot
    predict injury counts on zero distances. However, this should not happen
   as we should have total distances for all modes for the entire population (possibly inferred 
   from other modes) that appear in the injury data. - This data is used to parameterise the
   Poisson injury model.
   
 - A new list (injuries_list) is created containing all strike and casualty mode and age and sex combinations 
   together with strike and casualty mode distances for the baseline and all scenarios. For 
   the whw model, any strike mode and casualty pairs where strike mode equals casualty mode
   are removed. Combinations which do not have either a zero strike or casualty mode distance 
   are also removed. - This list will later be used to predict fatality counts using the 
   Poisson injury model.

 - The casualty and strike mode exponents are added to both the injuries_for_model and injuries_list
 
 -	The best possible regression model is being built using Baseline injury counts and distances 
   (injuries_for_model) such that the standard errors are small wherever possible. Strike 
   and casualty mode pairs where cas mode = strike mode are removed if they still exist
   which they should not as they should have been removed by the ithim_load_data.R function.
   Two different forms for whw and nov matrices are defined, taking into account age and sex
   information where it exists. The standard errors of the newly built regression models 
   are checked and if they are too large and if the data has not been aggregated by age 
   and sex yet, then the data is aggregated and a new Poisson regression model is build. 
   If the standard errors are still large, then a message is printed to the screen warning
   that the standard errors are large.
}
