% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data_for_voi.R
\name{extract_data_for_voi}
\alias{extract_data_for_voi}
\title{Get ITHIM-results into correct format for VoI analysis}
\usage{
extract_data_for_voi(
  NSCEN,
  NSAMPLES,
  SCEN_SHORT_NAME,
  cities,
  multi_city_ithim,
  output_version,
  level1,
  level2,
  level3
)
}
\arguments{
\item{NSCEN}{number of scenarios (not incl. baseline)}

\item{NSAMPLES}{number of model runs per city}

\item{SCEN_SHORT_NAME}{names of the scenarios (incl. baseline)}

\item{cities}{list of cities for which the model was run}

\item{multi_city_ithim}{list containing the ithim model information including results for the various model runs}

\item{output_version}{the output version of the model run}
}
\value{
ithim_results list with the following objects:

voi_data_all_df: dataframe for all cities with all outcomes for all model runs, age and sex categories and disease and scenario combinations
}
\description{
This function extracts the relevant information from the multi_city_ithim object and gets the results into
the correct format for further analysis.
}
\details{
The function performs the following steps:

\itemize{
\item by looping through the cities:

 \itemize{
   \item calculate average outcome (yll) per person in the population considered by the model 
   \item calculate the total ylls per 100 000 for each outcome age category, scenario and disease combination and model run
   \item calculate total yll outcome across all outcome age and sex categories per city and scenario and disease combinations
   \item calculate total yll outcome across all outcome sex categories per city and scenario and disease combinations
   \item create one dataframe for all cities with all outcomes for all model runs, age groups and disease and 
         scenario combinations
 } 
  
\item compute yll per hundred thousand by outcome age group by summing across all diseases (double counting!) 
     by city and scenario and also summing across all cities

\item create one dateframe with total ylls (median, 5th and 95th percentiles) per age group and city 
     (plus combined results as sum across all cities)
}
}
