% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scenario_settings.R
\name{get_scenario_settings}
\alias{get_scenario_settings}
\title{Get values for max mode share scenario}
\usage{
get_scenario_settings(
  cities = c("accra", "bangalore", "belo_horizonte", "bogota", "buenos_aires",
    "cape_town", "delhi", "mexico_city", "santiago", "sao_paulo", "vizag"),
  modes = c("pedestrian", "cycle", "car", "motorcycle", "bus"),
  distances = c("0-2 km", "2-6 km", "6+ km"),
  speeds = list(bus = 10, bus_driver = 10, minibus = 10, minibus_driver = 10, car =
    14.4, car_driver = 14.4, taxi = 12.6, pedestrian = 2.5, walk_to_pt = 2.5, cycle =
    7.2, motorcycle = 15.2, truck = 10, van = 14.4, subway = 18.1, rail = 21.9,
    auto_rickshaw = 4, shared_auto = 14.4, shared_taxi = 12.6, cycle_rickshaw = 4, other
    = 9.1)
)
}
\arguments{
\item{cities}{which cities to use}

\item{modes}{which modes to use}

\item{distances}{which distance categories to use}

\item{speeds}{named list of mode speeds (to be applied to all cities)}
}
\value{
data frame of proportions by mode and distance category
}
\description{
Computes the maximum mode share for specified mode types and specified distance categories across specified (stored) cities.
Used for max mode share scenario generation.
}
