% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PA_dose_response.R
\name{PA_dose_response}
\alias{PA_dose_response}
\title{Calculate RR given PA}
\usage{
PA_dose_response(cause, dose, confidence_intervals = F)
}
\arguments{
\item{cause}{name of disease}

\item{dose}{vector of doses of PA from individuals}

\item{confidence_intervals}{logic: whether or not to return confidence intervals}
}
\value{
data frame of relative risks
}
\description{
Calculate RR for a disease given PA
}
