% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_evppi_for_AP.R
\name{parallel_evppi_for_AP}
\alias{parallel_evppi_for_AP}
\title{Compute AP EVPPI}
\usage{
parallel_evppi_for_AP(disease, parameter_samples, outcome, NSCEN)
}
\arguments{
\item{disease}{disease name}

\item{parameter_samples}{data frame of parameter samples}

\item{outcome}{data frame of outcomes}

\item{NSCEN}{number of scenarios}
}
\value{
vector of EVPPI values (one per scenario)
}
\description{
For use to compute AP EVPPI in parallel
}
