% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_attributable_fraction.R
\name{population_attributable_fraction}
\alias{population_attributable_fraction}
\title{Calculate population attributable fraction}
\usage{
population_attributable_fraction(pop, cn, mat)
}
\arguments{
\item{pop}{}

\item{cn}{}

\item{mat}{}
}
\value{
population attributable fractions by demographic group
}
\description{
Calculate population attributable fraction
}
