% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ithim.R
\name{run_ithim}
\alias{run_ithim}
\title{Wrapper for running ITHIM}
\usage{
run_ithim(ithim_object, seed = 1)
}
\arguments{
\item{ithim_object}{list of input data needed to calculate the health burden}

\item{seed}{random seed}
}
\value{
ithim_object list of items giving the input data and output results
}
\description{
Switch to either calculate the health burden using the constant input parameter values or to sample from distributions first
}
\details{
This function works by creating a switch to run the computation by calling
\code{\link{ithim_calculation_sequence()}} directly, or to divert to the sampling case
\code{\link{ithim_uncertainty()}} which first extracts the sampled parameters and then calls
the \code{\link{ithim_calculation_sequence()}}.
}
