% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_scenario_specific_variables.R
\name{set_scenario_specific_variables}
\alias{set_scenario_specific_variables}
\title{Set scenario specific variables - CURRENTLY NOT CALLED}
\usage{
set_scenario_specific_variables()
}
\description{
Set NSCEN, SCEN and SCEN_SHORT_NAME
}
