% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_vehicle_inventory.R
\name{set_vehicle_inventory}
\alias{set_vehicle_inventory}
\title{Collate all vehicle information}
\usage{
set_vehicle_inventory()
}
\description{
Puts all vehicle information in one place including speeds and emission factors. Writes to global environment.
}
\details{
This function performs the following operations:

\itemize{
\item Based on the city specific (or default) mode speeds, the vehicle inventory is initialised

\item PM emissions are added from the city specific (or default) PM inventory

\item big_truck mode is added for which there is no distance but it this mode is used in the air pollution modules

\item other mode is added if it does not already exist in the travel survey (and hence in the mode speeds)

\item CO2 emissions are added from the city specific (or default) CO2 inventory

\item if car_driver exists, the car_driver emissions are set to the PM and CO2 car emissions

\item VEHICLE_INVENTORY is set as a global variable
}
}
