% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_ithim_inputs.R
\name{summarise_ithim_inputs}
\alias{summarise_ithim_inputs}
\title{Graphical processing of input data}
\usage{
summarise_ithim_inputs(ithim_object)
}
\arguments{
\item{ithim_object}{processed ithim_object from run_ithim_setup}
}
\description{
Produce graphs summarising some input data, e.g. travel, injury, AP
}
