% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_glm_object.R
\name{trim_glm_object}
\alias{trim_glm_object}
\title{Reduce size of glm object}
\usage{
trim_glm_object(obj)
}
\arguments{
\item{obj}{glm object}
}
\value{
glm object
}
\description{
Delete some attributes of glm object in order to save space
}
