% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_distance_columns.R
\name{add_distance_columns}
\alias{add_distance_columns}
\title{Add distance columns to injury tables}
\usage{
add_distance_columns(
  injury_table,
  mode_names,
  true_distances_0,
  dist,
  scenarios = SCEN
)
}
\arguments{
\item{injury_table}{(list of) data frame(s) to be edited}

\item{mode_names}{which modes to take distances for}

\item{true_distances_0}{distances to add to injury table}

\item{dist}{table used to access bus distance}

\item{scenarios}{which scenarios to process}
}
\value{
edited (list of) data frame(s)
}
\description{
Add distance columns to injury tables, matching on information in the injury contingency table
}
