% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ghost_trips.R
\name{add_ghost_trips}
\alias{add_ghost_trips}
\title{Add trips taken by no one}
\usage{
add_ghost_trips(
  raw_trip_set,
  trip_mode = "bus_driver",
  distance_ratio = BUS_TO_PASSENGER_RATIO * DISTANCE_SCALAR_PT,
  reference_mode = "bus"
)
}
\arguments{
\item{raw_trip_set}{data frame of trips}

\item{trip_mode}{which mode to add}

\item{distance_ratio}{fraction of reference distance to create}

\item{reference_mode}{name of reference mode}
}
\value{
data frame of trips
}
\description{
Creates trips based on distance relative to another (reference) mode, without altering the synthetic population
}
