% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trips.R
\name{add_trips}
\alias{add_trips}
\title{Add trips to trip set}
\usage{
add_trips(
  trip_ids = 0,
  new_mode = "pedestrian",
  distance = 1,
  participant_id = 0,
  age = 20,
  sex = "male",
  nTrips = 3,
  speed = 4.8
)
}
\arguments{
\item{trip_ids}{ids for new trips}

\item{new_mode}{mode for new trips}

\item{distance}{distances to sample from}

\item{participant_id}{participant id for new trips}

\item{age}{age for participant}

\item{sex}{sex for participant}

\item{nTrips}{number of trips for participant}

\item{speed}{speed for new trips}
}
\value{
data frame of trips
}
\description{
Creates a data frame of the same description as a trip set to append
}
