% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_walk_trips.R
\name{add_walk_trips}
\alias{add_walk_trips}
\title{Add pedestrian trips to trip set}
\usage{
add_walk_trips(pt_trips)
}
\arguments{
\item{pt_trips}{data frame of PT trips}
}
\value{
list of data frames of PT trips and walk-to-PT trips
}
\description{
Create data frame of walk-to-PT trips from PT trips and walk-to-bus time
}
