% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AP_dose_response.R
\name{AP_dose_response}
\alias{AP_dose_response}
\title{Calculate RR given AP}
\usage{
AP_dose_response(cause, dose, quantile, confidence_intervals = F)
}
\arguments{
\item{cause}{name of disease}

\item{dose}{vector of doses of AP from individuals}

\item{confidence_intervals}{logic: whether or not to return confidence intervals}
}
\value{
data frame of relative risks
}
\description{
Calculate RR for a disease given AP
}
