% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_age_groups.R
\name{assign_age_groups}
\alias{assign_age_groups}
\title{Assign age groups to individuals}
\usage{
assign_age_groups(
  dataset,
  age_category = AGE_CATEGORY,
  age_lower_bounds = AGE_LOWER_BOUNDS,
  max_age = MAX_AGE,
  min_age = AGE_LOWER_BOUNDS[1],
  age_label = "age"
)
}
\arguments{
\item{dataset}{data frame to be edited}

\item{age_category}{vector of strings giving age categories}

\item{age_lower_bounds}{lower boundaries of age categories}

\item{max_age}{maximum age for model}

\item{min_age}{minimum age for model}

\item{age_label}{string label for age column}
}
\value{
edited data frame
}
\description{
Prunes dataset given max and min ages; assigns age group labels given age
}
