% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_evppi.R
\name{call_evppi}
\alias{call_evppi}
\title{EVPPI wrapper function for total population YLLs}
\usage{
call_evppi(
  parameter_samples,
  outcome_voi_list,
  voi_complete_df,
  cities,
  NSCEN,
  NSAMPLES,
  SCEN_SHORT_NAME,
  scenario_names,
  output_version,
  level1,
  level2,
  level3
)
}
\arguments{
\item{parameter_samples}{table containing all the input parameter variables for the different model runs for all cities}

\item{outcome_voi_list}{vector detailing the outcomes to be considered in the VoI analysis}

\item{voi_complete_df}{total yll outcome for all outcome age categories per city and scenario and disease combination, also combined city result (sum)}

\item{cities}{vector of cities}

\item{NSCEN}{number of scenarios (not incl. baseline)}

\item{NSAMPLES}{number of times the model was run for each city}

\item{scenario_names}{gives the names of the scenarios (incl baseline)}

\item{output_version}{output version number}

\item{level1}{list of outcomes included in level 1}

\item{level2}{list of outcomes included in level 2}

\item{level3}{list of outcomes included in level 3}
}
\value{
list containg the following elements:

evppi_df containing the evppi values for all input parameters and scenario and disease outcomes

evppi_outcome_names, a vector containing all the scenario outcome combinations
}
\description{
This function gets the ithim results into the correct format and calls the \code{\link{compute_evppi()}} script
to calculate the EVPPIs for all input parameters and required outcomes and scenarios for the entire
population considered in the model
}
\details{
The function performs the following steps:

\itemize{
\item create a vector containing the global parameters but not including the dose response quantiles as they are treated separately
\item loop through the cities:
 \itemize{
   \item extract the RR values for each dose response function and outcome, calculate the RR for the 
         different levels and for the entire population
   \item extract the city specific parameters (excluding the CO2 and PM emission inventory parameters as they are not independent of each other)
   \item extract the outcomes of interest for each scenario using the outcome_voi_list
   \item call the \code{\link{compute_evppi()}} function to calculate the expected values of partially perfect information (EVPPI)
         for all parameters and diseases of interest
   \item if NSAMPLES >= 1000 then also calculate the EVPPI values for the emission inventory parameters
   } 
\item tidy up results data and add parameter classification
}
}
