% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_evppi_age_sex.R
\name{call_evppi_age_sex}
\alias{call_evppi_age_sex}
\title{EVPPI wrapper function for total population YLLs by sex and age}
\usage{
call_evppi_age_sex(
  voi_data_all_df,
  parameter_samples,
  outcome_voi_list,
  outcome,
  cities,
  voi_add_sum,
  NSCEN,
  NSAMPLES,
  scenario_names,
  evppi_df
)
}
\arguments{
\item{voi_data_all_df}{dataframe containing all outcomes by age and sex}

\item{parameter_samples}{table containing all the input parameter variables for the different model runs for all cities}

\item{outcome_voi_list}{vector detailing the outcomes to be considered in the VoI analysis}

\item{outcome}{total yll outcome for all outcome age categories per city and scenario and disease combination, also combined city result (sum)}

\item{cities}{vector of cities}

\item{voi_add_sum}{if the sum of YLLs across all disease outcomes is to be considered}

\item{NSCEN}{number of scenarios (not incl. baseline)}

\item{NSAMPLES}{number of times the model was run for each city}

\item{scenario_names}{gives the names of the scenarios (incl baseline)}

\item{evppi_df}{outcome dataframe containing voi analysis for total yll across all age and sex categories}
}
\value{
evppi_agesex_df dataframe containing all EVPPI outcomes for all age and sex categories and all cities

age_gender_cat vector with all age and sex categories

evppi_city_list_all list where each list entry is a dataframe containing all EvPPI outcomes for all age and sex categories for one city
}
\description{
This function gets the ithim results into the correct format and calls the \code{\link{compute_evppi()}} script
to calculate the EVPPIs for all input parameters and required outcomes and scenarios split by sex and age.
}
\details{
The function performs the following steps:

\itemize{
\item create a vector containing the global parameters but not including the emission inventory 
     parameters as they are not independent of each other

\item loop through the cities:
 \itemize{
   \item extract the city specific parameters (excluding the CO2 and PM emission inventory parameters)
   \item loop through all age and sex categories:
     \itemize{   
       \item extract the outcomes of interest for each scenario using the outcome_voi_list
       \item if voi_add_sum == TRUE, calculate the total YLLs by summing across all 
         diseases in the outcome_voi_list - this only makes sense if the diseases
         in the outcome_voi_list are independent of each other
       \item call the \code{\link{compute_evppi()}} function to calculate the expected values of partially perfect information (EVPPI)
         for all parameters and diseases of interest
       }
   \item if NSAMPLES >= 1000 then also calculate the EVPPI values for the emission inventory parameters by looping through 
         all age and sex categories
   } 
}
}
