% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_rr_ap_pa.R
\name{combined_rr_ap_pa}
\alias{combined_rr_ap_pa}
\title{Combine relative risks from air pollution and physical activity}
\usage{
combined_rr_ap_pa(ind_pa, ind_ap, conf_int = FALSE)
}
\arguments{
\item{ind_pa}{data frame of individual RRs for diseases affected by PA}

\item{ind_ap}{data frame of individual RRs for diseases affected by AP}

\item{conf_int}{logic: whether to include confidence interval from dose response relationships or not}
}
\value{
dataframe giving the RR risk for AP, PA and combined AP and PA exposure levels for every person in the synthetic population and for each scenario
}
\description{
Combine relative risks (RR) from air pollution (AP) and physical activity (PA) 
through multiplication for diseases affected
by both AP and PA
}
\details{
This function performs the following steps:

\itemize{
\item join the ap and pa relative risk datasets 

\item loop through all disease outcomes that are affected by both PA and AP:
    \itemize{
    \item for each scenario multiply the relative risks for PA and AP and store in a new column

    \item if confidence intervals are required, multiply the upper and lower RR for AP
    and PA respectively wherever possible, otherwise use the given median RR values instead
    }
}
}
