% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_trip_distance_duration.R
\name{complete_trip_distance_duration}
\alias{complete_trip_distance_duration}
\title{Add missing trip information}
\usage{
complete_trip_distance_duration()
}
\description{
Adds any missing trip information such as stage duration or stage distance.
}
\details{
This function performs to following steps to calculate stage duration or stage distance. Note that
we need at least trip duration or stage distance or stage stage duration:

\itemize{
\item add speed column

\item if there is no stage duration:
 \itemize{
 \item calculate the stage duration if the stage distance is given using the mode speeds

 \item if the stage duration sum is 'NA' (i.e. if not all stage distances exist or if there are no stage distances at all):
     \itemize{
     \item if trip duration exists and its sum is not NA, then find the trip ids with 'NA' stage duration and replace those
       stage duration with the trip duration divided by the number of stages of that particular trip

     \item remove any remaining trips with 'NA' stage distance

     \item update the stage speed column
     }
   }

\item if there is no stage distance:
  \itemize{
  \item calculate stage distance using the stage duration and mode speeds
   }
\item if there is no trip distance:
 \itemize{
 \item calculate trip distance by summing over the stage distances
}
}
}
