% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_trip_distance_duration.R
\name{complete_trip_distance_duration}
\alias{complete_trip_distance_duration}
\title{Data harmonisation: add columns to trip set if missing}
\usage{
complete_trip_distance_duration()
}
\description{
Creates any columns needed and missing from the trip set, then writes the trip set to the global enviroment
}
