% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_evppi.R
\name{compute_evppi}
\alias{compute_evppi}
\title{Compute evppi, designed to be run in parallel - ONLY USED in SAMPLING MODE}
\usage{
compute_evppi(p, global_para, city_para, city_outcomes, nsamples)
}
\arguments{
\item{p}{input parameter index}

\item{global_para}{list of global input parameters that are the same across all cities}

\item{city_para}{list of city specific input parameters}

\item{city_outcome}{list of outcomes for a specific city}
}
\value{
list of EVPPI vectors for specific city
}
\description{
Creates a list of EVPPI values one parameter (set) at a time across some pre-defined outcomes
Uses Chris Jackson's VoI Github package https://github.com/chjackson/voi
}
