% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_evppi.R
\name{compute_evppi}
\alias{compute_evppi}
\title{Compute evppi, designed to be run in parallel}
\usage{
compute_evppi(jj, sources, outcome, nscen = 1, all = F, multi_city_outcome = T)
}
\arguments{
\item{jj}{index for input and (potentially) outcome}

\item{sources}{list of inputs}

\item{outcome}{list of outcomes}

\item{nscen}{number of scenarios (which constitute columns of outcome)}

\item{all}{whether or not to compute EVPPI for source against all outcomes. Default=F, so sources[[jj]] relates to outcome[[jj]]}

\item{multi_city_outcome}{whether the last outcome corresponds to all cities combine}
}
\value{
list of EVPPI vectors
}
\description{
Creates a list of EVPPI values one parameter (set) at a time across one or more outcomes
}
