% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_africa_india_scenarios.R
\name{create_africa_india_scenarios}
\alias{create_africa_india_scenarios}
\title{Create scenarios for African and Indian cities - NOT CURRENTLY USED}
\usage{
create_africa_india_scenarios(trip_set)
}
\arguments{
\item{trip_set}{data frame, baseline scenario}
}
\value{
list of baseline scenario and four mode scenarios
}
\description{
Creates three scenarios where in each one, the mode share of a given mode is elevated by a set
percentage of the total trips. The scenario-modes are cycle, car, and bus and motorcycle.
}
\details{
Add 5% of trips overall in such a way that the average mean mode share for each mode across the
three distance bands is preserved.
}
