% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bogota_scenarios.R
\name{create_bogota_scenarios}
\alias{create_bogota_scenarios}
\title{Create scenarios based on mode shares across different distance bands}
\usage{
create_bogota_scenarios(trip_set)
}
\arguments{
\item{trip_set}{data frame, baseline scenario}
}
\value{
list of baseline scenario and four mode scenarios
}
\description{
Creates three scenarios where, in each one, the mode share is elevated to the
certain values in each distance band. The scenario-modes are cycle, car, and bus.
}
\details{
Based on create_global_scenarios

Add scenario_increase proportion of trips overall in such a way that the average mode share for each mode across the
three distance bands is preserved.

https://www.dropbox.com/home/ITHIM%20Global/Methods%20and%20Processes/ScenarioDefn/BogotaScenario
}
