% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_global_scenarios.R
\name{create_global_scenarios}
\alias{create_global_scenarios}
\title{Create scenarios based on difference between median and max propensities for all cities}
\usage{
create_global_scenarios(trip_set)
}
\arguments{
\item{trip_set}{data frame, baseline scenario}
}
\value{
list of baseline scenario and four mode scenarios
}
\description{
Creates four scenarios where, in each one, the mode share is elevated to the
certain values in each distance band. The scenario-modes are cycle, car, and bus and motorcycle.
}
\details{
Based on create_latam_scenarios

Assume that each of the three distance categories gets approximately 5% of trips more, however, split those 15% for 
each mode by the distance band distribution for the median mode shares across all cities
https://www.dropbox.com/home/ITHIM%20Global/Methods%20and%20Processes/ScenarioDefn/GlobalScenario
}
