% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_latam_scenarios.R
\name{create_latam_scenarios}
\alias{create_latam_scenarios}
\title{Create scenarios defined for latam paper}
\usage{
create_latam_scenarios(trip_set)
}
\arguments{
\item{trip_set}{data frame, baseline scenario}
}
\value{
list of baseline scenario and three mode scenarios
}
\description{
Creates three scenarios where, in each one, the mode share is elevated to the
certain values in each distance band. The scenario-modes are cycle, car, and bus.
}
\details{
Motorcycle trips were added as well
}
