% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_synth_pop.R
\name{create_synth_pop}
\alias{create_synth_pop}
\title{Creates synthetic population}
\usage{
create_synth_pop(raw_trip_set)
}
\arguments{
\item{raw_trip_set}{data frame of raw trips taken}
}
\value{
the synthetic population and the trip set which has been pruned
}
\description{
Creates a synthetic population by matching individuals in the trip set to individuals in the PA set
}
