% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_dur_tbls.R
\name{dist_dur_tbls}
\alias{dist_dur_tbls}
\title{Get distances and duration summaries by mode}
\usage{
dist_dur_tbls(trip_scen_sets)
}
\arguments{
\item{trip_scen_sets}{list of synthetic trip sets for each scenario including the baseline}
}
\value{
list of table of (total) distances and durations per mode per scenario
}
\description{
Summaries of total distances and durations spent travelling per mode and per scenario, for the synthetic population
}
\details{
The function performs the following steps:

\itemize{
\item loop through the scenarios (incl Baseline):
 \itemize{
  \item using the trip data, sum across the distances by stage mode to get total
    distance by mode for each scenario for the synthetic population
    
  \item sum across the duration by stage mode to get total duration by mode for each scenario
  
  \item if 'walk_to_pt' stages exist, add them to the pedestrian stages for both 
    duration and distance
  }  
\item create one dataframe containing the total distances by mode for each scenario

\item create one dataframe containing the total duration by mode for each scenario

\item remove any 'walk_to_pt' stages as they have been added to the pedestrian stages

\item update the bus and car driver distances and duration in the scenarios using 
  the ratio of bus/car to bus_driver/car_driver in the baseline scenario (this
  is redundant for the GLOBAL, BOGOTA, AFRICA_INDIA and LATAM scenario definitions
  as bus and car driver distances are already updated during the scenario creation)
}
}
