% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_dur_tbls.R
\name{dist_dur_tbls}
\alias{dist_dur_tbls}
\title{Get distances and durations}
\usage{
dist_dur_tbls(trip_scen_sets)
}
\arguments{
\item{trip_scen_sets}{list of synthetic trip sets for scenarios}
}
\value{
list of table of (total) distances and durations per mode per scenario
}
\description{
Summaries of total distances and durations spent travelling per mode per scenario
}
