% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_for_injury_function.R
\name{distances_for_injury_function}
\alias{distances_for_injury_function}
\title{Get distances and model for injuries module}
\usage{
distances_for_injury_function(journeys, dist)
}
\arguments{
\item{journeys}{df with sex and age_cat total distance (by total population) for scenarios}

\item{dist}{table of (total) distances per mode per scenario}
}
\value{
list of distances, injury table, and injury regression model
}
\description{
Computes exposures (distances) to parametrise the injury regression model, 
which is computed as a Poisson with various offsets and used later in prediction
}
