% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ap_rr.R
\name{gen_ap_rr}
\alias{gen_ap_rr}
\title{Get RR for diseases given AP exposure}
\usage{
gen_ap_rr(pm_conc_pp)
}
\arguments{
\item{pm_conc_pp}{individual AP exposures}
}
\value{
data frame of relative risks per person per disease
}
\description{
Computes the RR for individuals for each disease given AP exposure
}
