% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pa_rr.R
\name{gen_pa_rr}
\alias{gen_pa_rr}
\title{Get RR for diseases given mMETs}
\usage{
gen_pa_rr(mmets_pp)
}
\arguments{
\item{mmets_pp}{individual mMETs}
}
\value{
data frame of relative risks per person per disease
}
\description{
Computes the RR for individuals for each disease given mMETs (PA exposure)
}
