% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_distances.R
\name{get_all_distances}
\alias{get_all_distances}
\title{Sequence to get distance data}
\usage{
get_all_distances(ithim_object)
}
\arguments{
\item{ithim_object}{list containing synthetic trip set}
}
\value{
ithim_object again, with additional distance objects
}
\description{
Sequence of function calls to get distance data for modules from synthetic population
}
