% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scenario_settings.R
\name{get_scenario_settings}
\alias{get_scenario_settings}
\title{Get values for max mode share scenario}
\usage{
get_scenario_settings(
  cities = c("accra", "sao_paulo", "delhi", "bangalore", "santiago", "belo_horizonte",
    "buenos_aires", "mexico_city", "bogota", "cape_town", "vizag"),
  modes = c("pedestrian", "cycle", "car", "motorcycle", "bus"),
  distances = c("0-1 km", "2-5 km", "6+ km"),
  speeds = list(bus = 15, bus_driver = 15, minibus = 15, minibus_driver = 15, car = 21,
    taxi = 21, pedestrian = 4.8, walk_to_pt = 4.8, cycle = 14.5, motorcycle = 25, truck =
    21, van = 15, subway = 28, rail = 35, auto_rickshaw = 22, shared_auto = 22,
    cycle_rickshaw = 10)
)
}
\arguments{
\item{cities}{which cities to use}

\item{modes}{which modes to use}

\item{distances}{which distance categories to use}

\item{speeds}{named list of mode speeds (to be applied to all cities)}
}
\value{
data frame of proportions by mode and distance category
}
\description{
Computes the maximum mode share for specified mode types and specified distance categories across specified (stored) cities.
Used for max mode share scenario generation.
}
