% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_synthetic_from_trips.R
\name{get_synthetic_from_trips}
\alias{get_synthetic_from_trips}
\title{Generate synthetic population from trip data}
\usage{
get_synthetic_from_trips()
}
\value{
data frame of all synthetic trips from all scenarios
}
\description{
Sequence of functions to set up the synthetic population, the synthetic trips, and the scenarios.
}
\details{
This function performs the following steps:

\itemize{
\item The columns from the TRIP_SET are put into the correct order

\item multiply the trip distances, stage distances, and durations by the day_to_week scalar
  and then divide by 7 to get the distances and durations of an 'average' day of the week

\item add bus_driver and truck trips if required (\code{\link{add_ghost_trips()}})

\item add personal motorcycle trips if needed (call the appropriate function)

\item add commercial motorcycle trips if required (\code{\link{add_ghost_trips()}})

\item build the synthetic population by creating a data set that contains the (non-zero) participant
  ids and demographic information from the trip data set and adds work and leisure MMET
  values by calling \code{\link{create_synth_pop()}} (non travel entries in the trip data set are also removed)

\item adds car driver trips if required (\code{\link{add_ghost_trips()}})

\item call the \code{\link{ithim_setup_baseline_scenario()}} function to get the baseline data into the correct format
  for the creation of the different scenarios

\item create the required scenarios by calling the appropriate function

\item add walk to pt trips and combine the scenarios into one dataframe by calling the
  \code{\link{walk_to_pt_and_combine_scen()}} function
}
}
