% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_synthetic_from_trips.R
\name{get_synthetic_from_trips}
\alias{get_synthetic_from_trips}
\title{Generate synthetic data from trip set}
\usage{
get_synthetic_from_trips()
}
\value{
data frame of all trips from all scenarios
}
\description{
Sequence of functions to set up the synthetic population, the synthetic trips, and the scenarios.
Also sets global variables for later use.
}
