% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_burden.R
\name{health_burden}
\alias{health_burden}
\title{Compute health burden}
\usage{
health_burden(ind_ap_pa, combined_AP_PA = T)
}
\arguments{
\item{ind_ap_pa}{data.frame of all individuals' relative risks for diseases}

\item{combined_AP_PA=T}{logic: whether to combine the two exposure pathways (AP and PA) or to compute independently}
}
\value{
list of data.frames: one for deaths per disease per demographic group, and likewise for YLLs
}
\description{
Compute health burden for populations in scenarios given relative risks
}
