% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injuries_function_2.R
\name{injuries_function_2}
\alias{injuries_function_2}
\title{Predict injuries}
\usage{
injuries_function_2(
  true_distances,
  injuries_list,
  reg_model,
  constant_mode = F
)
}
\arguments{
\item{true_distances}{data frame to set up results}

\item{injuries_list}{list of data frames to supply to regression model for prediction}

\item{reg_model}{regression glm object}

\item{constant_mode}{whether or not we are in constant (vs sampling) mode}
}
\value{
list of injury prediction data frames and whw matrices
}
\description{
Predict injuries based on regression model from baseline and scenario travel
}
