% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injury_death_to_yll.R
\name{injury_death_to_yll}
\alias{injury_death_to_yll}
\title{Map injury death burden to YLL burden}
\usage{
injury_death_to_yll(injuries)
}
\arguments{
\item{injuries}{data frame of injury deaths}
}
\value{
list of injury deaths and YLLs (which are differences from reference scenario) plus the values in the reference scenario.
}
\description{
Calculated the YLL burden from the death burden of injury based on the ratio in the GBD data.
}
