% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ithim.R
\name{ithim_calculation_sequence}
\alias{ithim_calculation_sequence}
\title{Cascade of computations that form ITHIM-Global}
\usage{
ithim_calculation_sequence(ithim_object, seed = 1)
}
\arguments{
\item{ithim_object}{name of disease}

\item{seed}{random seed}
}
\value{
ithim_object - list of items making up the ithim result
}
\description{
Ordered set of computations as part of ITHIM-Global that calculates the required output parameters
}
\details{
This function performs the following steps:

\enumerate{

\item extract all lists and variables from the ithim_object list

\item  air pollution pathway:
   \itemize{
   \item calculate the PM2.5 exposure for each person in the baseline population
     and PM2.5 emissions for each mode and scenario (\code{\link{scenario_pm_calculations()}})

   \item calculate the CO2 emissions for each mode and scenario (\code{\link{scenario_co2_calculations()}})

   \item assign relative risk to each person in the baseline population for each disease
     related to PM pollution and each scenario based on the individual PM exposure
     levels (\code{\link{gen_ap_rr()}})
     }

\item  physical activity pathway:
   \itemize{
   \item calculate total mMETs for each person in the baseline population (\code{\link{total_mmet()}})

   \item assign relative risk to each person in the baseline population for each disease
     related to physical activity levels and each scenario based on the individual mMET
     values (\code{\link{gen_pa_rr()}})
   }

\item  physical activity and air pollution combined:
   \itemize{
   \item combine the PA and AP datasets by joining the two datasets. For disease affected by
     both PA and AP calculate the joined relative risk by multiplying the PA and AP
     relative risks (\code{\link{combined_rr_ap_pa()}})

   \item calculate the health burden (Yll and deaths) for each disease and age and sex
     category (\code{\link{health_burden()}}):
     \itemize{
     \item calculate the health burden (Yll and deaths) for each disease and age
       and sex category. Combine the AP and PA pathways for diseases affected
       by both AP and PA

     \item if running in constant mode also calculate the health burden for both the
       AP and PA pathways separately
       }
    }


\item  combine all pathways using the outputs from 3. and 4.:
   \itemize{
   \item combine the AP, PA and health burden data for ylls and deaths
     for all diseases and scenarios
     }
}
}
