% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ithim.R
\name{ithim_calculation_sequence}
\alias{ithim_calculation_sequence}
\title{Cascade of computations that form the ITHIM}
\usage{
ithim_calculation_sequence(ithim_object, seed = 1)
}
\arguments{
\item{ithim_object}{name of disease}

\item{seed}{}
}
\value{
list of items making up the ithim result
}
\description{
Ordered set of computations that form the ITHIM, from travel information to health burden.
}
