% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ithim_load_data.R
\name{ithim_load_data}
\alias{ithim_load_data}
\title{Load data for model and setting}
\usage{
ithim_load_data(
  speeds = list(bus = 10, bus_driver = 10, minibus = 10, minibus_driver = 10, car =
    14.4, car_driver = 14.4, taxi = 12.6, pedestrian = 2.5, walk_to_pt = 2.5, cycle =
    7.2, motorcycle = 15.2, truck = 10, van = 14.4, subway = 18.1, rail = 21.9,
    auto_rickshaw = 4, shared_auto = 14.4, shared_taxi = 12.6, cycle_rickshaw = 4, other
    = 9.1)
)
}
\arguments{
\item{speeds}{named list of mode speeds}
}
\description{
Loads and processes data from file. Local data for the setting and global data for the model.
Writes objects to the global environment.
}
\details{
# @param setup_call_summary_filename name of file to write to - removed this parameter
}
