% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ithim_load_data.R
\name{ithim_load_data}
\alias{ithim_load_data}
\title{Load data for model and setting}
\usage{
ithim_load_data(
  setup_call_summary_filename,
  speeds = list(bus = 11, bus_driver = 11, minibus = 11, minibus_driver = 11, car = 14,
    taxi = 9, pedestrian = 3, walk_to_pt = 3, cycle = 7, motorcycle = 14, truck = 11, van
    = 14, subway = 16, rail = 20, auto_rickshaw = 9, shared_auto = 14, shared_taxi = 9,
    cycle_rickshaw = 5)
)
}
\arguments{
\item{setup_call_summary_filename}{name of file to write to}

\item{speeds}{named list of mode speeds}
}
\description{
Loads and processes data from file. Local data for the setting and global data for the model.
Writes objects to the global environment.
}
