% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ithim_setup_baseline_scenario.R
\name{ithim_setup_baseline_scenario}
\alias{ithim_setup_baseline_scenario}
\title{Set up baseline scenario data frame}
\usage{
ithim_setup_baseline_scenario(trip_set)
}
\arguments{
\item{trip_set}{data frame of trips}
}
\value{
trip_set as baseline scenario
}
\description{
Create scenario by adding distance categories and scenario=baseline column to trip set data frame
}
