% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ithim_uncertainty.R
\name{ithim_uncertainty}
\alias{ithim_uncertainty}
\title{Sampling routine for running ITHIM with uncertainty}
\usage{
ithim_uncertainty(ithim_object, seed = 1)
}
\arguments{
\item{ithim_object}{list of necessary inputs, including parameters}

\item{seed}{which sample to take}
}
\value{
list of ITHIM outcomes
}
\description{
Sets sampled parameters to the global environment and calls the ITHIM routine
}
\details{
This function works by performing the following steps:

\itemize{
\item extract the ithim_object list entries  

\item extract the sampled parameters which are stored in the ithim_object  
\item call set_vehicle_inventory.R to update emissions if any  
  emission parameters have been sampled from a distribution   
  
\item call get_synthetic_from_trips.R to update the synthetic trips and synthetic population if any 
  relevant input parameters have been sampled from a distribution 
   
\item call get_all_distances.R to recalculate the distances if any of the relevant input parameters have
  been sampled from a distribution  
  
\item run ITHIM-Global model by calling ithim_calculation_sequence.R
}
}
