% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ithim_uncertainty.R
\name{ithim_uncertainty}
\alias{ithim_uncertainty}
\title{Sampling routine for calculating ITHIM with uncertainty}
\usage{
ithim_uncertainty(ithim_object, seed = 1)
}
\arguments{
\item{ithim_object}{list of necessary inputs, including parameters}

\item{seed}{which sample to take}
}
\value{
list of ITHIM outcomes
}
\description{
Sets sampled parameters to the global environment and runs the ITHIM routine (run_ithim)
}
