% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ithim.R
\name{run_ithim}
\alias{run_ithim}
\title{Wrapper for running ITHIM}
\usage{
run_ithim(ithim_object, seed = 1)
}
\arguments{
\item{ithim_object}{list of items making up the ithim set up}

\item{seed}{}
}
\value{
ithim_object list of items making up the ithim result
}
\description{
Switch to run the computation directly, or divert to the sampling case
}
