% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ithim_setup.R
\name{run_ithim_setup}
\alias{run_ithim_setup}
\title{Run the set up script for ITHIM}
\usage{
run_ithim_setup(
  seed = 1,
  CITY = "accra",
  speeds = NULL,
  PM_emission_inventory = NULL,
  CO2_emission_inventory = NULL,
  DIST_CAT = c("0-6 km", "7-9 km", "10+ km"),
  AGE_RANGE = c(0, 150),
  ADD_WALK_TO_PT_TRIPS = T,
  ADD_BUS_DRIVERS = T,
  ADD_CAR_DRIVERS = T,
  ADD_TRUCK_DRIVERS = T,
  ADD_MOTORCYCLE_FLEET = F,
  ADD_PERSONAL_MOTORCYCLE_TRIPS = "no",
  TEST_WALK_SCENARIO = F,
  TEST_CYCLE_SCENARIO = F,
  MAX_MODE_SHARE_SCENARIO = F,
  REFERENCE_SCENARIO = "Baseline",
  PATH_TO_LOCAL_DATA = NULL,
  NSAMPLES = 1,
  BUS_WALK_TIME = 5,
  RAIL_WALK_TIME = 15,
  MMET_CYCLING = 4.63,
  MMET_WALKING = 2.53,
  PM_CONC_BASE = 50,
  PM_TRANS_SHARE = 0.225,
  PA_DOSE_RESPONSE_QUANTILE = F,
  AP_DOSE_RESPONSE_QUANTILE = F,
  BACKGROUND_PA_SCALAR = 1,
  BACKGROUND_PA_CONFIDENCE = 1,
  INJURY_REPORTING_RATE = 1,
  CHRONIC_DISEASE_SCALAR = 1,
  DAY_TO_WEEK_TRAVEL_SCALAR = 7,
  SIN_EXPONENT_SUM = 2,
  CASUALTY_EXPONENT_FRACTION = 0.5,
  SIN_EXPONENT_SUM_NOV = 1,
  SIN_EXPONENT_SUM_CYCLE = 2,
  CASUALTY_EXPONENT_FRACTION_CYCLE = 0.5,
  SIN_EXPONENT_SUM_PED = 2,
  CASUALTY_EXPONENT_FRACTION_PED = 0.5,
  SIN_EXPONENT_SUM_VEH = 2,
  CASUALTY_EXPONENT_FRACTION_VEH = 0.5,
  CALL_INDIVIDUAL_SIN = F,
  BUS_TO_PASSENGER_RATIO = 0.022,
  CAR_OCCUPANCY_RATIO = 0.6,
  TRUCK_TO_CAR_RATIO = 0.21,
  FLEET_TO_MOTORCYCLE_RATIO = 0.01,
  PROPORTION_MOTORCYCLE_TRIPS = 0,
  PM_EMISSION_INVENTORY_CONFIDENCE = 1,
  CO2_EMISSION_INVENTORY_CONFIDENCE = 1,
  DISTANCE_SCALAR_CAR_TAXI = 1,
  DISTANCE_SCALAR_WALKING = 1,
  DISTANCE_SCALAR_PT = 1,
  DISTANCE_SCALAR_CYCLING = 1,
  DISTANCE_SCALAR_MOTORCYCLE = 1,
  LATAM = FALSE,
  GLOBAL = FALSE
)
}
\arguments{
\item{seed}{random seed}

\item{CITY}{name of the city, and name of the directory containing city data files}

\item{speeds}{named list of mode speeds}

\item{PM_emission_inventory}{named list of mode emissions}

\item{CO2_emission_inventory}{named list of mode emissions}

\item{DIST_CAT}{vector string of distance categories in the form '0-6'. (The unit is assumed to be the same as in the trip set.)}

\item{AGE_RANGE}{vector of minimum and maximum ages to include}

\item{ADD_WALK_TO_PT_TRIPS}{logic: whether or not to add short walks to all PT trips}

\item{ADD_BUS_DRIVERS}{logic: whether or not to add bus drivers}

\item{ADD_CAR_DRIVERS}{logic: whether or not to find and add distance travelled by individual cars, denoted by car drivers}

\item{ADD_TRUCK_DRIVERS}{logic: whether or not to add truck drivers}

\item{ADD_MOTORCYCLE_FLEET}{logic: whether or not to add additional commercial motorcycle fleet as ghost trips}

\item{ADD_PERSONAL_MOTORCYCLE_TRIPS}{character: if 'no' does not add motorcycle trips otherwise set to geographic region which defines the set-up of the motorcycle trips to be added}

\item{TEST_WALK_SCENARIO}{logic: whether or not to run the walk scenario}

\item{TEST_CYCLE_SCENARIO}{logic: whether or not to run the cycle scenario}

\item{MAX_MODE_SHARE_SCENARIO}{logic: whether or not to run the max mode share scenario}

\item{REFERENCE_SCENARIO}{which scenario forms the reference for the health comparison}

\item{PATH_TO_LOCAL_DATA}{path to CITY directory, if not using package}

\item{NSAMPLES}{constant integer: number of samples to take}

\item{BUS_WALK_TIME}{lognormal parameter: duration of walk to Bus}

\item{RAIL_WALK_TIME}{lognormal parameter: duration of walk to Rail}

\item{MMET_CYCLING}{lognormal parameter: mMETs when cycling}

\item{MMET_WALKING}{lognormal parameter: mMETs when walking}

\item{PM_CONC_BASE}{lognormal parameter: background PM2.5 concentration}

\item{PM_TRANS_SHARE}{beta parameter: fraction of background PM2.5 attributable to transport}

\item{PA_DOSE_RESPONSE_QUANTILE}{logic: whether or not to sample from PA RR DR functions}

\item{AP_DOSE_RESPONSE_QUANTILE}{logic: whether or not to sample from AP RR DR functions}

\item{BACKGROUND_PA_SCALAR}{lognormal parameter: reporting scalar for PA}

\item{BACKGROUND_PA_CONFIDENCE}{beta parameter: confidence in accuracy of PA survey}

\item{INJURY_REPORTING_RATE}{lognormal parameter: rate of injury reporting}

\item{CHRONIC_DISEASE_SCALAR}{lognormal parameter: scalar for background disease rates}

\item{DAY_TO_WEEK_TRAVEL_SCALAR}{beta parameter: rate of scaling travel from one day to one week}

\item{SIN_EXPONENT_SUM}{lognormal parameter: linearity of injuries with respect to two modes. SIN_EXPONENT_SUM=2 means no safety in numbers.}

\item{CASUALTY_EXPONENT_FRACTION}{beta parameter: casualty contribution to SIN_EXPONENT_SUM}

\item{SIN_EXPONENT_SUM_NOV}{lognormal parameter: linearity of injuries with respect to two modes where strike mode = NOV. SIN_EXPONENT_SUM=2 means no safety in numbers.}

\item{SIN_EXPONENT_SUM_CYCLE}{lognormal parameter: linearity of injuries with respect to two modes where victim mode = cycle. SIN_EXPONENT_SUM=2 means no safety in numbers.}

\item{CASUALTY_EXPONENT_FRACTION_CYCLE}{beta parameter: casualty contribution to SIN_EXPONENT_SUM  where victim mode = cycle.}

\item{SIN_EXPONENT_SUM_PED}{lognormal parameter: linearity of injuries with respect to two modes  where victim mode = pedestrian. SIN_EXPONENT_SUM=2 means no safety in numbers.}

\item{CASUALTY_EXPONENT_FRACTION_PED}{beta parameter: casualty contribution to SIN_EXPONENT_SUM where victim mode = pedestrian}

\item{SIN_EXPONENT_SUM_VEH}{lognormal parameter: linearity of injuries with respect to two modes where victim mode = a vehicle. SIN_EXPONENT_SUM=2 means no safety in numbers.}

\item{CASUALTY_EXPONENT_FRACTION_VEH}{beta parameter: casualty contribution to SIN_EXPONENT_SUM where victim mode = a vehicle.}

\item{CALL_INDIVIDUAL_SIN}{logic: whether or not to call the safety in number coefficients for individual vehicles or use the same coefficients for all modes.}

\item{BUS_TO_PASSENGER_RATIO}{beta parameter: number of buses per passenger}

\item{CAR_OCCUPANCY_RATIO}{beta parameter: number of people per car (including driver)}

\item{TRUCK_TO_CAR_RATIO}{beta parameter: number of trucks per car}

\item{FLEET_TO_MOTORCYCLE_RATIO}{beta parameter: fraction of total motorcycles that's fleet}

\item{PROPORTION_MOTORCYCLE_TRIPS}{beta parameter: proportion of trips that are to be added as motorcycle trips}

\item{PM_EMISSION_INVENTORY_CONFIDENCE}{beta parameter: confidence in accuracy of emission inventory}

\item{CO2_EMISSION_INVENTORY_CONFIDENCE}{beta parameter: confidence in accuracy of emission inventory}

\item{DISTANCE_SCALAR_CAR_TAXI}{lognormal parameter: scalar for car distance travelled}

\item{DISTANCE_SCALAR_WALKING}{lognormal parameter: scalar for walking distance travelled}

\item{DISTANCE_SCALAR_PT}{lognormal parameter: scalar for PT distance travelled}

\item{DISTANCE_SCALAR_CYCLING}{lognormal parameter: scalar for cycling distance travelled}

\item{DISTANCE_SCALAR_MOTORCYCLE}{lognormal parameter: scalar for motorcycle distance travelled}

\item{setup_call_summary_filename}{name to write setup call summary to}
}
\value{
ithim_object list of objects for onward use.
}
\description{
Sets up the basic ITHIM object for onward calculation. Data loading, processing and harmonisation. Setting global values.
}
\details{
Parameters have two options: to be set to a constant, and to be sampled from a prespecified distribution.
Each parameter is given as an argument of length 1 or 2. 
If length 1, it's constant, and set to the global environment. 
If length 2, a distribution is defined and sampled from NSAMPLE times.
There are some exceptions, listed above
}
