% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk_to_pt_and_combine_scen.R
\name{scale_trip_distances}
\alias{scale_trip_distances}
\title{Scale trip distances}
\usage{
scale_trip_distances(trips)
}
\arguments{
\item{trips}{data frame, all trips from all scenarios}
}
\value{
data frame, all trips from all scenarios
}
\description{
Applies mode-specific distance scalars to all trips
}
