% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_co2_calculations.R
\name{scenario_co2_calculations}
\alias{scenario_co2_calculations}
\title{Calculate total CO2 exposure per mode and scenario}
\usage{
scenario_co2_calculations(dist)
}
\arguments{
\item{dist}{data frame of population travel from all scenarios}
}
\value{
total CO2 exposure per mode
}
\description{
Calculate total CO2 exposure for each mode and for each scenario based on the CO2 emissions inventory
}
\details{
This function performs the following steps:

\itemize{
\item calculate emission factors for each mode by dividing total emissions by distances travelled

\item calculate CO2 emissions for each mode in each scenario by multiplying the scenario distance times the emission factors

\item for modes without any assigned distance, use the CO2 emissions from the VEHICLE_INVENTORY instead
}
}
