% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_co2_calculations.R
\name{scenario_co2_calculations}
\alias{scenario_co2_calculations}
\title{Calculate total CO2 exposure per mode}
\usage{
scenario_co2_calculations(dist)
}
\arguments{
\item{dist}{data frame of population travel from all scenarios}
}
\value{
total CO2 exposure per mode
}
\description{
Calculate total CO2 exposure per person based on population and personal travel
}
