% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_pm_calculations - Copy.R,
%   R/scenario_pm_calculations.R
\name{scenario_pm_calculations}
\alias{scenario_pm_calculations}
\title{Calculate total AP exposure per person}
\usage{
scenario_pm_calculations(dist, trip_scen_sets)

scenario_pm_calculations(dist, trip_scen_sets)
}
\arguments{
\item{dist}{total distance travelled by mode by the population for all scenarios}

\item{trip_scen_sets}{trips data frame of all trips from all scenarios}
}
\value{
background PM concentration for baseline and all scenarios

total AP exposure per person in the synthetic population (for baseline and scenarios)

background PM concentration for baseline and all scenarios

total AP exposure per person in the synthetic population (for baseline and scenarios)
}
\description{
Calculate total AP exposure per person based on population and personal travel

Calculate total AP exposure per person based on population and personal travel
}
\details{
This function performs the following steps:

\itemize{
\item the ventilation rates per mode are defined - these parameters are fixed

\item the exposure factor rates by activity are defined - these parameters are fixed

\item calculate pm concentration not related to transport

\item calculate PM emission factors for each mode by dividing total emissions by distances travelled

\item calculate PM emissions for each mode in each scenario by multiplying the scenario distance
  by the emission factors

\item for modes without any assigned distance, use the PM emissions from the VEHICLE_INVENTORY instead

\item calculate the total PM concentrations for each scenario

\item add ventilation and exposure factors to the trip set by stage mode

\item add total scenario PM concentrations to the trip set

\item add the inhaled air and total PM (in micro grams) to the trip set

\item define the amount of time per day spent as leisure sedentary screen time, 
  non-discretionary time and other time - fixed

\item add total time spent travelling by each participant to the trip set

\item calculate the sleep and the rest ventilation rates

\item for each participant in the synthetic population (with travel component), 
  calculate the total air inhaled, the total PM inhaled and 
  the total PM concentration inhaled for each scenario

\item assign all participants in the synthetic population without travel component,
  the baseline or scenario PM concentrations

\item join all people with and without travel in the synthetic population
}

This function performs the following steps:

\itemize{
\item the exposure factor rates by activity are defined - these parameters are fixed

\item calculate pm concentration not related to transport

\item calculate PM emission factors for each mode by dividing total emissions by distances travelled

\item calculate PM emissions for each mode in each scenario by multiplying the scenario distance
  by the emission factors

\item for modes without any assigned distance, use the PM emissions from the VEHICLE_INVENTORY instead

\item calculate the total PM concentrations for each scenario

\item add exposure factors to the trip set by stage mode

\item add total scenario PM concentrations to the trip set

\item calculate ventilation rate for each stage taking into account demographic characteristics and exposure factors

\item calculate the inhaled air and total PM (in micro grams) in the trip set

\item calculate the amount of time per day spent in sleep, moderate and vigorous
  activities

\item add total time spent travelling by each participant to the trip set

\item calculate ventilation rate for sleep, moderate and vigorous activities

\item for each participant in the synthetic population (with travel component),
  calculate the total air inhaled, the total PM inhaled and
  the total PM concentration inhaled for each scenario

\item assign all participants in the synthetic population without travel component,
  the baseline or scenario PM concentrations

\item join all people with and without travel in the synthetic population
}
}
