% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_pm_calculations.R
\name{scenario_pm_calculations}
\alias{scenario_pm_calculations}
\title{Calculate total AP exposure per person}
\usage{
scenario_pm_calculations(dist, trip_scen_sets)
}
\arguments{
\item{dist}{data frame of population travel from all scenarios}

\item{trip_scen_sets}{data frame of all trips from all scenarios}
}
\value{
background AP

total AP exposure per person
}
\description{
Calculate total AP exposure per person based on population and personal travel
}
