% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_injury_contingency.R
\name{set_injury_contingency}
\alias{set_injury_contingency}
\title{Create contingency table from itemised list of injuries}
\usage{
set_injury_contingency(injuries)
}
\arguments{
\item{injuries}{data frame of injury events}
}
\description{
One of the inputs is a list of injury events. This function aggregates injuries by type into a long contingency table with prespecified column names.
Write tables to global environment.
}
