% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_vehicle_inventory.R
\name{set_vehicle_inventory}
\alias{set_vehicle_inventory}
\title{Collate all vehicle information}
\usage{
set_vehicle_inventory()
}
\description{
Puts all vehicle information in one place. Writes to global environment.
}
