% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_mmet.R
\name{total_mmet}
\alias{total_mmet}
\title{Calculate total mMETs per person}
\usage{
total_mmet(trip_scen_sets)
}
\arguments{
\item{trip_scen_sets}{data frame of all trips from all scenarios}
}
\value{
total mMETs per week per person
}
\description{
Calculate total mMETs per person based on PA and active travel
}
