% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk_to_pt_and_combine_scen.R
\name{walk_to_pt_and_combine_scen}
\alias{walk_to_pt_and_combine_scen}
\title{Add walk to public transport stages}
\usage{
walk_to_pt_and_combine_scen(SYNTHETIC_TRIPS)
}
\arguments{
\item{trip_set}{list of data frames, trips from all scenarios}
}
\value{
data frame, all trips from all scenarios
}
\description{
Adds a short walk stage to any public transport (PT) trip if required.
}
\details{
This function performs the following steps:

\itemize{
\item create a list containing the dataframes of the synthetic trips for each scenario

\item if ADD_WALK_TO_PT_TRIPS == T, i.e if additional 'walk to pt' stages are to be added:
  \itemize{
  \item filter out all trips with a public transport stage mode

  \item divide public transport trips into those with and without a 'walk to pt' stage

 \item add a 'walk to pt' stage to those public transport trips without a walking stage
    (\code{\link{add_walk_trips()}})
    }
\item combine all trips from all scenarios into one dataframe

\item scale the stage distances and durations by calling the scale_trip_distances.R function
}
}
